Clarify the scope of these standards in the top-level readme. (#14)
Sam's attempt at a clear and charming summary of the reasoning for
publishing a style guide. Picky comments about language are very
welcome. I wanted to be a bit more brief, but I couldn't think of a way.

### Note

In the meeting we actually said we "should" use them unless there was a
good reason. But "should not" change for existing projects.

### Also

Emoji's in a public-facing site: yay or nay? It's very on-brand for ARC.
So I went with yay.