Add C++ style guidelines. (#15)
Includes and expands on [our internal
documentation](https://github.com/UCL-ARC/research-software-documentation/blob/main/processes/programming_projects/group_practices.md#coding-conventions).

~I stuck with Google style because that was in our internal docs and
it's what I know and like. If we have a consensus to switch to...
something else... I'm fine with that.~
_**Edit**: style tweaked based on discussion_ 👇

My opinion is: that we should have a decision that most of us agree
with. I don't actually have a very strong opinion about which style to
choose.

From the [many many possible
tools](https://github.com/UCL-ARC/coding-standards/issues/6#issuecomment-1772483519)
I have stuck to a single linter (`clang-tidy`) which is also what we
discuss in PHAS0100, and `cppcheck` as a single static analyser. Perhaps
obviously, I want to avoid tools that no one in ARC has used or has
experience with. We can add to them if we find something that is
life-changing.

@mmcleod89: following our IRL chat a couple of weeks ago, there's now
some explanation of these guidelines' scope in the [top-level
README](https://github.com/UCL-ARC/coding-standards#readme). Anyone who
often works in C++ and has an opinion is welcome to join the review. The
starting reviewers are C++istas that I've spoken to about this.

## Solves

* #6

---------

Co-authored-by: Christian Gutschow <64213771+20DM@users.noreply.github.com>
Co-authored-by: Milan Malfait <m.malfait@ucl.ac.uk>